object dependencies changes and integrated into single wrapper
